Removed models, added JPA entities instead

+++ server/src/main/java/com/dzalex/skillshuffle/controllers/ChatController.java
@@ -3,7 +3,7 @@ package com.dzalex.skillshuffle.controllers;
 import com.dzalex.skillshuffle.dtos.ChatDTO;
 import com.dzalex.skillshuffle.dtos.ChatPreviewDTO;
 import com.dzalex.skillshuffle.dtos.MessageDTO;
-import com.dzalex.skillshuffle.models.Chat;
+import com.dzalex.skillshuffle.entities.Chat;
 import com.dzalex.skillshuffle.repositories.ChatRepository;
 import com.dzalex.skillshuffle.services.ChatService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,13 +26,13 @@ public class ChatController {
     }
 
     @GetMapping("/chats/{id}")
-    public ChatDTO getChat(@PathVariable("id") Long id) {
+    public ChatDTO getChat(@PathVariable("id") Integer id) {
         Chat chat = chatRepository.findChatById(id);
         return chatService.getChatWithMessages(chat);
     }
 
     @GetMapping("/chats/{id}/messages")
-    public List<MessageDTO> getChatMessages(@PathVariable("id") Long id,
+    public List<MessageDTO> getChatMessages(@PathVariable("id") Integer id,
                                             @RequestParam(defaultValue = "30") int limit,
                                             @RequestParam(defaultValue = "0") int offset) {
         return chatService.getChatMessages(id, limit, offset);
