+++ server/src/main/java/com/dzalex/skillshuffle/repositories/ChatMemberRepository.java
@@ -1,10 +1,10 @@
 package com.dzalex.skillshuffle.repositories;
 
-import com.dzalex.skillshuffle.models.ChatMember;
+import com.dzalex.skillshuffle.entities.ChatMember;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
 public interface ChatMemberRepository extends JpaRepository<ChatMember, Long> {
-    List<ChatMember> findAllByChatId(Long chatId);
+    List<ChatMember> findAllByChatId(Integer chatId);
 }
